service/requirements.txt
asgiref==3.2.7
astroid==2.3.3
boto3==1.13.19
botocore==1.16.19
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
Django==3.0.6
djangorestframework==3.11.0
docutils==0.15.2
dodgy==0.2.1
Faker==4.1.0
idna==2.9
isort==4.3.21
jmespath==0.10.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.4
pep8-naming==0.4.1
pika==1.1.0
psycopg2==2.8.5
pycodestyle==2.4.0
pydocstyle==5.0.2
pyflakes==2.1.1
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.12
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requirements-detector==0.7
s3transfer==0.3.3
setoptconf==0.2.0
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
wrapt==1.11.2
